Add files via upload
perhaps not the most elegant code ever written but 100% my work.  I created this code that takes in a message then adds a random letter to every other character then reverses it, and (using a randomly generated alphabet) changes every letter to its correlation on the random alphabet.  This code uses the random seed to ensure it can be decrypted later.